+++ sync-for-payables/CodatSyncPayables/Models/Operations/CreateConnectionRequestBody.cs
@@ -8,7 +8,7 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatSyncPayables.Models.Operations
+namespace Codat.Sync.Payables.Models.Operations
 {
     using Newtonsoft.Json;