+++ sync-for-payables/CodatSyncPayables/Models/Operations/ListTaxRatesResponse.cs
@@ -8,9 +8,9 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatSyncPayables.Models.Operations
+namespace Codat.Sync.Payables.Models.Operations
 {
-    using CodatSyncPayables.Models.Shared;
+    using Codat.Sync.Payables.Models.Shared;
     using System.Net.Http;
     using System;
     
@@ -35,11 +35,11 @@ namespace CodatSyncPayables.Models.Operations
         /// <summary>
         /// Raw HTTP response; suitable for custom response parsing
         /// </summary>
-        public HttpResponseMessage? RawResponse { get; set; }
+        public HttpResponseMessage RawResponse { get; set; } = default!;
 
         /// <summary>
         /// Success
         /// </summary>
-        public TaxRates? TaxRates { get; set; }
+        public Models.Shared.TaxRates? TaxRates { get; set; }
     }
 }